refactor(s2n-quic-platform): simplify Message trait
